// my-app/src/app/models/users-model.ts
export class UsersModel {    
    id: number;
    firstName: string;
    lastName: string;
    userName: string;
    password: string;
    newPassword: string;
    transactionPassword: string;
    newTransactionPassword: string;
    holderName: string;
    accountNumber: string;
    cretedOn: any;
    contactNumber: string;
    email: string;
    lastLogin: any;
    location: string;
    gender: string;
    dateOfBirth: any;
    accountbalance:any;
    randomText:string;
}